Moved calculate_gradient_across_cell_faces.
Moved calculate_gradient_across_cell_faces to the raster_gradients
module.
